Web App/requirements.txt
absl-py==0.11.0
appdirs==1.4.4
astunparse==1.6.3
audioread==2.1.9
beautifulsoup4==4.9.3
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
Cython==0.29.14
decorator==4.4.2
ffmpeg-python==0.2.0
Flask==1.1.2
future==0.18.2
gast==0.3.3
gensim==3.8.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.34.0
h5py==2.10.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
Keras-Preprocessing==1.1.2
librosa==0.8.0
llvmlite==0.35.0
Markdown==3.3.3
MarkupSafe==1.1.1
norbert==0.2.1
numba==0.52.0
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.7
pandas==1.1.2
pooch==1.3.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydub==0.24.1
pyparsing==2.4.7
python-crfsuite==0.9.7
python-dateutil==2.8.1
pytz==2020.4
pyvi==0.1
requests==2.25.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.6
scikit-learn==0.23.2
scipy==1.4.1
six==1.15.0
sklearn-crfsuite==0.3.6
smart-open==4.0.1
SoundFile==0.10.3.post1
soupsieve==2.1
SpeechRecognition==3.8.1
spleeter==2.0.2
tabulate==0.8.7
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.54.1
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
